@@ -12,7 +12,11 @@
     "access": "public"
   },
   "scripts": {
-    "precommit": "lint-staged"
+    "clean": "rimraf lib",
+    "precommit": "lint-staged",
+    "prebuild": "npm run clean",
+    "prepare": "npm run build",
+    "build": "tsc -p ."
   },
   "repository": {
     "url": "tinajs/mina-webpack",
@@ -23,24 +27,35 @@
   },
   "homepage": "https://github.com/tinajs/mina-webpack/tree/master/packages/mina-entry-webpack-plugin#readme",
   "peerDependencies": {
-    "@tinajs/mina-loader": "^1.0.0-0",
-    "webpack": "^4.29.6"
+    "@tinajs/mina-loader": "^1.5.1",
+    "webpack": "^4.32.2"
   },
   "devDependencies": {
     "@tinajs/mina-loader": "^1.5.1",
+    "@types/compose-function": "^0.0.30",
+    "@types/fs-extra": "^7.0.0",
+    "@types/json5": "^0.0.30",
+    "@types/loader-utils": "^1.1.3",
+    "@types/minimatch": "^3.0.3",
+    "@types/p-props": "^1.0.2",
+    "@types/replace-ext": "^0.0.27",
+    "@types/resolve": "0.0.8",
+    "@types/webpack": "^4.4.32",
     "lint-staged": "^8.1.4",
     "prettier": "^1.16.4",
-    "webpack": "^4.29.6"
+    "typescript": "^3.3.3333",
+    "vue": "^2.6.10",
+    "webpack": "^4.32.2"
   },
   "dependencies": {
     "compose-function": "^3.0.3",
     "ensure-posix-path": "^1.1.1",
     "flatten": "^1.0.2",
-    "fs-extra": "^7.0.1",
+    "fs-extra": "^8.0.1",
     "json5": "^2.1.0",
     "loader-utils": "^1.2.3",
     "minimatch": "^3.0.4",
-    "p-props": "^1.2.0",
+    "p-props": "^3.0.0",
     "replace-ext": "^1.0.0",
     "resolve": "^1.10.0",
     "resolve-from": "^5.0.0",
