@@ -57,11 +57,11 @@
     "prettier": "^1.16.4",
     "ts-loader": "^5.3.3",
     "typescript": "^3.3.3333",
-    "webpack": "^4.29.6",
+    "webpack": "^4.32.2",
     "webpack-merge": "^4.2.1"
   },
   "peerDependencies": {
-    "webpack": "^4.0.0"
+    "webpack": "^4.32.2"
   },
   "dependencies": {
     "@tinajs/mina-sfc": "^0.2.2",
