@@ -1,10 +1,10 @@
-const fs = require('fs')
-const JSON5 = require('json5')
-const { parseComponent } = require('vue-template-compiler')
-const ConfigReader = require('../interfaces/config-reader')
+import fs from 'fs'
+import JSON5 from 'json5'
+import { parseComponent } from 'vue-template-compiler'
+import ConfigReader from '../interfaces/config-reader'
 
-class MinaConfigReader extends ConfigReader {
-  static getConfig(filePath) {
+export default class MinaConfigReader extends ConfigReader {
+  static getConfig(filePath: string) {
     let plain = fs.readFileSync(filePath, 'utf8')
     let blocks = parseComponent(plain).customBlocks
     let matched = blocks.find(block => block.type === 'config')
@@ -14,5 +14,3 @@ class MinaConfigReader extends ConfigReader {
     return JSON5.parse(matched.content)
   }
 }
-
-module.exports = MinaConfigReader
